Experiments-main/AI-Science-Lab/backend/requirements.txt
streamlit
sentence-transformers
faiss-cpu
pdfplumber
pymupdf
openai
numpy
scikit-learn
